+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # 1.1.6 - In progress
 
+- [Feature] Custom hex keyboard
 - [Feature] Add NFC Editor with colors and card
 - [Feature] Support Unknown firmware
 - [Feature] Compose Navigation on "Info" Screen